+++ pkgs/client/src/App.tsx
@@ -5,22 +5,22 @@ import Auth from './pages/Auth';
 import Dashboard from './pages/Dashboard';
 
 import './App.css';
-import Recipients from './pages/Recipients';
 import Invoice from './pages/Invoice';
 import User from './pages/User';
 import NewRecipients from './pages/Recipients/New Recipient';
+import UpdateRecipients from './pages/Recipients/Update Recipient';
 
 function App() {
   return (
     <StrictMode>
       <BrowserRouter>
         <Routes>
-          <Route path="/" element={<Dashboard />} />
-          <Route path="/auth" element={<Auth />} />
-          <Route path="/recipients" element={<Recipients />} />
-          <Route path="/new-recipients" element={<NewRecipients />} />
-          <Route path="/invoice" element={<Invoice />} />
-          <Route path="/user" element={<User />} />
+          <Route path="/dashboard" element={ <Dashboard /> } />
+          <Route path="/auth" element={ <Auth /> } />
+          <Route path="/update-recipient" element={ <UpdateRecipients /> } />
+          <Route path="/new-recipients" element={ <NewRecipients /> } />
+          <Route path="/invoice" element={ <Invoice /> } />
+          <Route path="/user" element={ <User /> } />
         </Routes>
       </BrowserRouter>
     </StrictMode>
