+++ pkgs/client/src/pages/Auth/Auth.css
@@ -7,3 +7,24 @@
     color:blueviolet;
     padding-left: 1%;
 }
+
+.auth-span{
+    display: flex;
+    flex-direction: column;
+    font-display: center;
+    place-items: center;
+}
+
+.page-title{
+  display: flex;
+  flex-direction: column;
+  place-items: center;
+}
+
+.error-message {
+  color: red;
+  margin-top: 0.5rem;
+  margin-bottom: 1rem;
+  font-size: 1rem;
+  font-family: inherit;
+}
