--- pkgs/client/src/pages/Auth/components/SignUp/SignUp.css
@@ -1,15 +1,0 @@
-.sign-up_form {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    padding: 1rem;
-    outline-color: blueviolet;
-}
-
-.error-message {
-  color: red;
-  margin-top: 0.5rem;
-  margin-bottom: 1rem;
-  font-size: 1rem;
-  font-family: inherit;
-}
