+++ pkgs/client/src/pages/Dashboard/index.tsx
@@ -1,23 +1,65 @@
-import { Link } from 'react-router-dom';
+import { useEffect, useState } from 'react';
 
 import Navbar from '../Auth/components/Navbar';
+import EmptyState from './components/EmptyState';
+import UserDashboard from './components/UserDashboard';
+import { SendInvoiceClient } from '../../services/SendInvoice';
 
 import './Dashboard.css';
+import type { Invoice } from '../../services/SendInvoice/Invoice';
 
 export default function Dashboard() {
+    const [hasInvoices, setHasInvoices] = useState(false);     
+    const [isLoading, setIsLoading] = useState(true);                  
+    const [invoices, setInvoices] = useState<Invoice[]>([]);
+    useEffect(() => {
+        const checkUserInvoices = async () => {
+            try {
+                const sendInvoiceClient = new SendInvoiceClient('http://127.0.0.1:8080' as any);
+                const invoiceList = await sendInvoiceClient.invoice.getInvoices();
+                debugger;
+                
+                if (!invoiceList || invoiceList.length === 0) {
+                    setHasInvoices(false);
+                    setInvoices([]);
+                } else {
+                    setHasInvoices(true);
+                    setInvoices(invoiceList);
+                }
+
+                setIsLoading(false);
+            } catch (error) {
+                console.error('Error loading invoices:', error);
+                setHasInvoices(false);
+                setInvoices([]);
+                setIsLoading(false);
+            }
+        };
+
+        checkUserInvoices();
+    }, []);
+
+    if (isLoading) {
+        return (
+            <div className="dashboard-layout">
+                <Navbar />
+                <div className="dashboard-content-area">
+                    <p>Loading...</p>
+                </div>
+            </div>
+        );
+    }
+
     return (
         <div className="dashboard-layout">
             <Navbar />
-            <div className="dashboard-content-area">
-                <h1 className="dashboard-title"> Welcome to Send Invoice</h1>
-                <h2 className="dashboard-subtitle"> Simplify your billing process. </h2>
-                <h3 className="dashboard-subtitle2">
-                    Easily create, manage, and send professional invoices to your clients in just a few clicks.
-                </h3>
-                <Link className='dashboard-link' to='/invoice' style={{ textDecoration: 'none' }}>
-                    <h3>Create New Invoce</h3>
-                </Link>
-            </div>
+            {
+                !hasInvoices ? (
+                    <EmptyState />
+                ) : (
+                    <UserDashboard invoices={invoices} />
+                )
+            }
         </div>
     );
 }
